MODIFY task_2.sql

+++ task_2.sql
@@ -39,7 +39,7 @@ CREATE TABLE IF NOT EXISTS Order_Details (
 orderdetailid INT UNSIGNED PRIMARY KEY AUTO_INCREMENT,
 order_id INT UNSIGNED NOT NULL,
 book_id INT UNSIGNED NOT NULL,
-quantity DECIMAL,
+quantity DOUBLE,
 
 FOREIGN KEY (order_id) REFERENCES Orders(order_id),
 FOREIGN KEY (book_id) REFERENCES Books(book_id)
